requirements.txt
python-telegram-bot
requests
flask
pyyaml
dataset